bot/requirements.txt
python-dotenv
pandas
numpy
chainlit==1.0.101
#langchain==0.0.354
langchain==0.1.0
langchain-community #==0.0.11
langchain-openai #==0.0.2
langchain-experimental
langchainhub
transformers
torch
openai
yfinance
wikipedia
duckduckgo-search
numexpr
